+++ src/rest/resources/DB/Database.java
@@ -2,6 +2,7 @@ package rest.resources.DB;
 
 import ofar.generated.classes.input.ServiceInput;
 
+import java.math.BigInteger;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -12,7 +13,7 @@ public class Database {
         if (serviceInput == null) {
             return null;
         }
-        serviceInput.setId(dbHashMap.size());
+        serviceInput.setId(BigInteger.valueOf(dbHashMap.size()));
         dbHashMap.put(dbHashMap.size(), serviceInput);
         return serviceInput;
     }
